create: replace vfsgen with embed
Since Go 1.16, it is possible to use "embed" library to work with
virtual file systems. We use one for templates in ``cartridge create``.
Before this patch, we used third party library [1]. Now it's replaced
with "embed".

To support embedding folder which only contents are file starting with
dot, we need to use Go 1.18. Moreover, file with "|" symbol in name
considered to be invalid [2] and its embedding is not supported.
So approach from #612 was reworked to use lowered app name as variable
instead of transforming it with Funcs.

Directories in vfs now have "dr-xr-xr-x" permissions instead of
"drwxr-xr-x", so we change it on template build (otherwise it would be
impossible to create subdirectories and files in directories).

Running ``go mod tidy`` regrouped go.mod contents, together with
cleaning up [1] dependency.

Closes #631

1. https://github.com/shurcooL/vfsgen
2. https://cs.opensource.google/go/x/mod/+/refs/tags/v0.5.1:module/module.go;l=488